backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.4
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.5
django-filter==23.3
django-import-export==3.3.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-extra-fields==3.7.0
et-xmlfile==1.1.0
filetype==1.2.0
gunicorn==21.2.0
idna==3.4
isort==5.12.0
MarkupPy==1.14
numpy==1.26.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
Pillow==10.0.1
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.5
xlrd==2.0.1
xlwt==1.3.0
